+++ SignalObj.cpp
@@ -55,7 +55,7 @@ void SignalObj::initTrItems(float* tpos){
     if(tpos == NULL)
         return;
     int trNodeId = tpos[0];
-    int metry = tpos[1];
+    float metry = tpos[1];
     
     TDB* tdb = Game::trackDB;
     qDebug() <<"new signal  "<<this->fileName;
